Refactor Register into AuthForm and use it for Register and Login, also change createUser for authUser
