+++ src/adapters/user.adapter.ts
@@ -1,6 +1,6 @@
 import { User } from "../models";
 
-export const createUserAdapter = (data: any): User => {
+export const authUserAdapter = (data: any): User => {
   return {
     id: data.data.user.id,
     username: data.data.user.username,
