+++ src/pages/index.ts
@@ -1,3 +1,4 @@
+export * from './AuthForm';
 export * from './Dashboard';
 export * from './HomePage';
 export * from './Login';
