+++ src/services/public.service.ts
@@ -17,3 +17,21 @@ export const register = (userData: FormValues) => {
     controller
   };
 };
+
+export const login = (userData: FormValues) => {
+  const controller = new AbortController();
+  return {
+    call: axios.post(
+      "http://127.0.0.1:8000/login",
+      userData,
+      {
+        headers: {
+          "Content-Type": "application/json",
+        },
+        signal: controller.signal
+      },
+    ),
+    controller
+  };
+};
+
